requirements.txt
retrying==1.3.3
cn2an==0.4.1
APScheduler==3.6.3
selenium
itchat-uos==1.5.0.dev
pandas
pyzbar==0.1.8
Pillow==8.3.2
suning-sdk==1.0.0
PyMySQL
PyYAML~=6.0.1
requests~=2.32.3